requirements.txt
numpy==1.19.5
pandas==1.0.1
plotly==5.1.0
streamlit==0.84.0
Pillow==7.0.0